+++ packages/web3-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @reef-knot/web3-react
 
+## 6.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/core-react@6.0.0
+
 ## 5.0.0
 
 ### Patch Changes